app/requirements.txt
mangum==0.17.0
fastapi==0.92.0
uvicorn==0.20.0
boto3
python-multipart
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
python-dotenv
